\name{GMMBoost-package}
\alias{GMMBoost-package}
\docType{package}
\title{
Likelihood-based Boosting for Generalized mixed models
}
\description{
Likelihood-based Boosting for Generalized mixed models
}
\details{
\tabular{ll}{
Package: \tab GMMBoost\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2012-03-29\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
for loading a dataset type data(nameofdataset)
}
\author{
Andreas Groll
}
\keyword{
Boosting, Variable selection, Generalized linear mixed model, Generalized additive mixed model
}
\seealso{
\code{\link{bGLMM},\link{OrdinalBoost},\link{bGAMM}}
}

\references{
Special thanks goes to Manuel Eugster, Sebastian Kaiser, Fabian Scheipl and Felix Heinzl, who helped to create this package and whose insightful advices helped to improve the package.
}


