\name{GMMBoost}
\alias{GMMBoost}
\docType{package}
\title{
Likelihood-Based Boosting for Generalized Mixed Models
}
\description{
This packages provides likelihood-based boosting approaches for Generalized mixed models
}
\details{
\tabular{ll}{
Package: \tab GMMBoost\cr
Type: \tab Package\cr
Version: \tab 1.1.3\cr
Date: \tab 2019-02-03\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
for loading a dataset type data(nameofdataset)
}
\author{
Andreas Groll
}
\concept{Boosting}
\concept{Variable selection}
\concept{Generalized linear mixed model}
\concept{Generalized additive mixed model}

\seealso{
\code{\link{bGLMM},\link{OrdinalBoost},\link{bGAMM}}
}

\references{
Special thanks goes to Manuel Eugster, Sebastian Kaiser, Fabian Scheipl and Felix Heinzl, who helped to create this package and whose insightful advices helped to improve the package.
}


